refactor: split EntityHelper into partials
