+++ pengdows.crud/EntityHelper.Caching.cs
@@ -1,0 +1,14 @@
+namespace pengdows.crud;
+
+public partial class EntityHelper<TEntity, TRowID>
+{
+    private const int MaxCacheSize = 100;
+
+    public void ClearCaches()
+    {
+        _readerPlans.Clear();
+        _columnListCache.Clear();
+        _queryCache.Clear();
+        _whereParameterNames.Clear();
+    }
+}
